+++ src/main/java/com/seyfi/review/model/request/CreateVoteDto.java
@@ -12,11 +12,11 @@ public class CreateVoteDto {
 
     @NotNull(message = "productId can't be null")
     @Positive(message = "productId should be a positive number")
-    private Integer productId;
+    private Long productId;
 
     @NotNull(message = "userId can't be null")
     @Positive(message = "userId should be a positive number")
-    private Integer userId;
+    private Long userId;
 
     @NotNull(message = "isCustomer can't be null")
     private Boolean isCustomer;
